requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.1
frozenlist==1.4.1
geoip2==4.8.0
idna==3.7
maxminddb==2.6.2
multidict==6.0.5
requests==2.32.3
setuptools==70.1.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
django-environ==0.11.2
django-redis==5.4.0
psycopg2-binary==2.9.9
djangorestframework==3.15.2
